third_party/blink/web_tests/http/tests/xmlhttprequest/async-xhr-revalidate-after-sync-xhr-expected.txt

Async XHR should not turn sync due to revalidation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xhr2SendIsReturned is true
PASS xhr1.responseText is not xhr2.responseText
PASS successfullyParsed is true

TEST COMPLETE
